payments modelling + endpoints

+++ events/models.py
@@ -72,7 +72,7 @@ class TicketType(BaseModel):
 class EventPromotion(BaseModel):
     name = models.CharField(max_length=256, null=False, blank=False)
     event = models.ForeignKey(Event, on_delete=models.CASCADE, null=False, blank=False)
-    promotion_rate = models.IntegerField(null=False, blank=False, default=10)
+    promotion_rate = models.FloatField(null=False, blank=False, default=10.0)
     expiry = models.DateField(null=False, blank=False, auto_now=False)
     use_limit = models.IntegerField(null=False, blank=False, default=100)
 
@@ -85,7 +85,7 @@ class TicketPromotion(BaseModel):
     ticket = models.ForeignKey(
         TicketType, on_delete=models.CASCADE, null=False, blank=False
     )
-    promotion_rate = models.IntegerField(null=False, blank=False, default=10)
+    promotion_rate = models.FloatField(null=False, blank=False, default=10.0)
     expiry = models.DateField(null=False, blank=False, auto_now=False)
     use_limit = models.IntegerField(null=False, blank=False, default=100)
